fix(6g): extract topology_only_snapshot test to root + pub test-access fields
CI Rust Compile (test) (--workspace --tests) caught two test-compile gaps in
the move that local `check --tests` (root-only) missed:

- The TD-168 `topology_only_snapshot_tests` mod (persistence.rs) used the root
  `crate::graph::unified_wal_factory()` + tempfile — like the td066 tests, it
  needs the root factory, so it can't live in the orion crate. Extracted to the
  root `orion_recovery_tests.rs` alongside td066 (rewrote its one
  `OrionPersistence::read_lock` CSR read to direct `csr_outgoing.read()`).
- The extracted assertions read engine internals now across the crate boundary:
  made `OrionGraphEngine.edge_metadata` + `node_to_index` and
  `CsrStorage.targets` `pub` (mirrors the `memory_pool`/`csr_outgoing` pub
  exposure already done for the td066 tests).

Verified: cargo check --workspace --lib --bins --tests (the CI gate) clean;
cargo clippy --lib --bins clean.
